opdrachten inlever

+++ 04 types/03 endless/Endless/Square.cs
@@ -5,7 +5,7 @@ namespace Endless
     internal class Square
     {
         //1) maak van int een byte
-        public int x, y;
+        public byte x, y;
         internal Brush color = Brushes.Red;
 
     }
